+++ CoinUtils/src/CoinParamUtils.cpp
@@ -304,7 +304,7 @@ int lookupParam (std::string name, CoinParamVec &paramVec,
   is, while 2 and 4 or more are queries about `?' or `???'.
 */
   int numQuery = 0 ;
-  { int length = name.length() ;
+  { int length = static_cast<int>(name.length()) ;
     int i ;
     for (i = length-1 ; i >= 0 && name[i] == '?' ; i--)
     { numQuery++ ; }
@@ -530,7 +530,7 @@ int matchParam (const CoinParamVec &paramVec, std::string name,
 		int &matchNdx, int &shortCnt)
 
 { 
-  int vecLen = paramVec.size() ;
+  int vecLen = static_cast<int>(paramVec.size()) ;
   int matchCnt = 0 ;
 
   matchNdx = -1 ;
@@ -566,7 +566,7 @@ int matchParam (const CoinParamVec &paramVec, std::string name,
 
 void printIt (const char *msg)
 
-{ int length = strlen(msg) ;
+{ int length = static_cast<int>(strlen(msg)) ;
   char temp[101] ;
   int i ;
   int n = 0 ;
@@ -598,7 +598,7 @@ void shortOrHelpOne (CoinParamVec &paramVec,
 		     int matchNdx, std::string name, int numQuery)
 
 { int i ;
-  int numParams = paramVec.size() ;
+  int numParams = static_cast<int>(paramVec.size()) ;
   int lclNdx = -1 ;
 /*
   For a short match, we need to look up the parameter again. This should find
@@ -656,7 +656,7 @@ void shortOrHelpOne (CoinParamVec &paramVec,
 
 void shortOrHelpMany (CoinParamVec &paramVec, std::string name, int numQuery)
 
-{ int numParams = paramVec.size() ;
+{ int numParams = static_cast<int>(paramVec.size()) ;
 /*
   Scan the parameter list. For each match, print just the name, or the name
   and short help.
@@ -669,7 +669,7 @@ void shortOrHelpMany (CoinParamVec &paramVec, std::string name, int numQuery)
     int match = param->matches(name) ;
     if (match > 0)
     { std::string nme = param->matchName() ;
-      int len = nme.length() ;
+      int len = static_cast<int>(nme.length()) ;
       if (numQuery >= 2) 
       { std::cout << nme << " : " << param->shortHelp() ;
 	std::cout << std::endl ; }
@@ -744,7 +744,7 @@ void printHelp (CoinParamVec &paramVec, int firstParam, int lastParam,
 
 { bool noHelp = !(shortHelp || longHelp) ;
   int i ;
-  int pfxLen = prefix.length() ;
+  int pfxLen = static_cast<int>(prefix.length()) ;
   bool printed = false ;
 
   if (noHelp)
@@ -754,7 +754,7 @@ void printHelp (CoinParamVec &paramVec, int firstParam, int lastParam,
       if (param == 0) continue ;
       if (param->display() || hidden)
       { std::string nme = param->matchName() ;
-	int len = nme.length() ;
+	int len = static_cast<int>(nme.length()) ;
 	if (!printed)
 	{ std::cout << std::endl << prefix ;
 	  lineLen += pfxLen ;
